commented responsive grid
commented base template

renamed css into convetion of 'grid' and 'tile'

a grid declares the intention for a set of simple tiles.

a tile can have it's width mutated in a later css include to
adjust how many columns exist in  the visual display of a device.

a grid begins a new column area.

by default there is a single column - more columns are created
by giving the tile a divisible size by the base wídth of the device
or media.